Refactor - components names and folders now have a pattern
