+++ src/components/LoginCover/elements/CoverText.js
@@ -1,0 +1,14 @@
+import styled from "styled-components";
+
+const CoverText = styled.h2`
+    font-family: Oswald;
+    font-size: 43px;
+    margin: 8px 0px;
+
+    @media (max-width: 937px) {
+        font-size: 23px;
+        margin: 5px 0px;
+    }
+`;
+
+export default CoverText;
