+++ src/components/Post/elements/Description.js
@@ -1,0 +1,10 @@
+import styled from "styled-components";
+
+const Description = styled.p`
+    font-size: 18px;
+    color: #B7B7B7;
+    display: inline-block;
+    margin: 8px 0px;
+`;
+
+export default Description;
