+++ src/components/Post/elements/IconButton.js
@@ -1,0 +1,11 @@
+import styled from "styled-components";
+
+const IconButton = styled.button`
+    font-size: 14px;
+    color: #FFFFFF;
+    position: absolute;
+    right: ${ ({right}) => right };
+    top: 0;
+`;
+
+export default IconButton;
