+++ src/components/Post/elements/LinkBox.js
@@ -1,0 +1,10 @@
+import styled from "styled-components";
+
+const LinkBox = styled.div`
+    width: 100%;
+    height: 155px;
+    border: 1px solid #C4C4C4;
+    border-radius: 11px;
+`;
+
+export default LinkBox;
