+++ src/pages/Timeline/Timeline.js
@@ -1,12 +1,14 @@
+import Loading from "../../components/Loading";
+import Container from "../../components/Container";
+import Header from "../../components/Header/Header";
+import PageTitle from "../../components/PageTitle";
+import Trending from "../../components/Trending";
+import Post from "../../components/Post/Post";
+
+import MockPosts from "../../temp_mocks/mock_posts";
+
 import { useState } from "react";
 import styled from "styled-components";
-import Container from "../../components/Container";
-import Trending from "../../components/Trending";
-import Post from "../../components/Post";
-import PageTitle from "../../components/PageTitle";
-import Loading from "../../components/Loading";
-import MockPosts from "../../temp_mocks/mock_posts";
-import Header from "../../components/Header";
 
 export default function Timeline() {
     const [posts] = useState(MockPosts.posts);
